# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "All HTTP + MQ endpoints exposed by the goop2 viewer.\\n\\nAll peer-to-peer signaling travels through MQ (POST /api/mq/send → P2P → /api/mq/events SSE).\\nSee x-mq-topics in the raw spec (/api/openapi.json) for the full topic contract.",
        "title": "goop2 Viewer API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/cluster/binary": {
            "post": {
                "description": "Sets the local binary that will execute jobs for this cluster. The binary is a child process that speaks the goop2 cluster JSON protocol over stdin/stdout. Mode can be \"oneshot\" (started per job) or \"daemon\" (long-running). Setting the binary resets the verified flag — the host must re-verify via check-job before dispatching work.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Set the binary path for this worker (worker only)",
                "parameters": [
                    {
                        "description": "Binary configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterBinaryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.clusterBinaryResponse"
                        }
                    },
                    "400": {
                        "description": "missing binary path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "not a cluster worker",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cluster/cancel": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Cancel a job (host only)",
                "parameters": [
                    {
                        "description": "Cancel request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterCancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "400": {
                        "description": "missing job_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "not a cluster host",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cluster/clear": {
            "post": {
                "description": "Removes all jobs (pending, completed, failed, cancelled) from the queue and database. This is irreversible.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Clear the entire job queue (host only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "409": {
                        "description": "not a cluster host",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cluster/create": {
            "post": {
                "description": "Creates a new cluster group and sets this node as host. If group_id is provided, activates an existing group instead of creating a new one.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Create or activate a cluster (become host)",
                "parameters": [
                    {
                        "description": "Create request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.clusterCreateResponse"
                        }
                    },
                    "409": {
                        "description": "already in a cluster",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cluster/delete": {
            "post": {
                "description": "Removes a cancelled, completed, or failed job from the queue and database. Active jobs (pending, assigned, running) cannot be deleted — cancel them first.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Delete a terminal job from the queue (host only)",
                "parameters": [
                    {
                        "description": "Delete request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "400": {
                        "description": "missing job_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "job is pending — cancel it first",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/cluster/join": {
            "post": {
                "description": "Joins a remote cluster group and registers this node as a worker.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Join an existing cluster as worker",
                "parameters": [
                    {
                        "description": "Join request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "400": {
                        "description": "missing host_peer_id or group_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "already in a cluster",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "failed to join group",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/cluster/pause": {
            "post": {
                "description": "Pauses the worker so the host dispatcher skips it when dispatching jobs. The worker stays in the cluster. Use POST /api/cluster/resume to resume.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Pause this worker (worker only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "409": {
                        "description": "not a cluster worker",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cluster/resume": {
            "post": {
                "description": "Resumes a paused worker so the host dispatcher can dispatch jobs to it again.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Resume this worker (worker only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "409": {
                        "description": "not a cluster worker",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/cluster/submit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Submit a job to the cluster queue (host only)",
                "parameters": [
                    {
                        "description": "Job submission",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterSubmitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.clusterSubmitResponse"
                        }
                    },
                    "400": {
                        "description": "missing job type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "not a cluster host",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/cluster/worker/pause": {
            "post": {
                "description": "Sends a pause command to the specified worker. The worker stops accepting new jobs but remains in the cluster. Existing running jobs are not interrupted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Pause a remote worker (host only)",
                "parameters": [
                    {
                        "description": "Worker peer ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterWorkerPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "400": {
                        "description": "missing peer_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "not a cluster host",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/cluster/worker/resume": {
            "post": {
                "description": "Sends a resume command to the specified worker so the dispatcher can dispatch jobs to it again.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Resume a remote worker (host only)",
                "parameters": [
                    {
                        "description": "Worker peer ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.clusterWorkerPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    },
                    "400": {
                        "description": "missing peer_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "not a cluster host",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.clusterBinaryRequest": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string",
                    "example": "oneshot"
                },
                "path": {
                    "type": "string",
                    "example": "/usr/bin/renderer"
                }
            }
        },
        "routes.clusterBinaryResponse": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string",
                    "example": "oneshot"
                },
                "path": {
                    "type": "string",
                    "example": "/usr/bin/renderer"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "routes.clusterCancelRequest": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "j-abc123"
                }
            }
        },
        "routes.clusterCreateRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "name": {
                    "type": "string",
                    "example": "My Cluster"
                }
            }
        },
        "routes.clusterCreateResponse": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "status": {
                    "type": "string",
                    "example": "created"
                }
            }
        },
        "routes.clusterDeleteRequest": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "j-abc123"
                }
            }
        },
...
        "routes.clusterJoinRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "host_peer_id": {
                    "type": "string",
                    "example": "12D3KooWXxx..."
                }
            }
        },
...
        "routes.clusterSubmitRequest": {
            "type": "object",
            "properties": {
                "max_retry": {
                    "type": "integer",
                    "example": 2
                },
                "mode": {
                    "type": "string",
                    "example": "oneshot"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "priority": {
                    "type": "integer",
                    "example": 5
                },
                "timeout_s": {
                    "type": "integer",
                    "example": 300
                },
                "type": {
                    "type": "string",
                    "example": "calculate"
                }
            }
        },
        "routes.clusterSubmitResponse": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "j-abc123"
                },
                "status": {
                    "type": "string",
                    "example": "submitted"
                }
            }
        },
...
        "routes.clusterWorkerPeerRequest": {
            "type": "object",
            "properties": {
                "peer_id": {
                    "type": "string",
                    "example": "12D3KooWXxx..."
                }
            }
        },
...
        "routes.statusOK": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    }
}